% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cid10_document.R
\docType{data}
\name{cid_aps}
\alias{cid_aps}
\title{Vetor com os códigos CID 10 de causas sensíveis a atenção primária}
\description{
Vetor com 560 Cids das condições sensíveis à atenção primária
}
\references{
\url{https://github.com/msrodrigues}
}
\author{
Marcio Rodrigues \email{msrodrigues@gmail.com}
}
\keyword{aps}
\keyword{cid10,}
\keyword{data,}
