% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cid10_document.R
\docType{data}
\name{cid_grupos}
\alias{cid_grupos}
\title{Tabela com os GRUPOS do CID 10}
\description{
Trata-se de um dataset de 275 linhas  e 4 colinas, com os agrupamentos de doenças.
}
\details{
\itemize{
      \item cat_ini: Código CID do inicio do grupo
      \item cat_fim: Codigo CID do fim do grupo
      \item descricao: Texto descrevendo a doença
      \item descricaoabrev: Texto abreviado descrevendo a doença
}
}
\examples{
cid_grupos

 A tibble: 275 x 4
 cat_ini cat_fim descricao                                    descrabrev
 <chr>   <chr>   <chr>                                        <chr>
 1 A00     A09     Doenças infecciosas intestinais              Doenças infecciosas intestinais
 2 A15     A19     Tuberculose                                  Tuberculose
 3 A20     A28     Algumas doenças bacterianas zoonóticas       Algumas doenças bacterianas zoonóticas
 4 A30     A49     Outras doenças bacterianas                   Outras doenças bacterianas
 5 A50     A64     Infecções de transmissão predominantemente … Infecções de transmissão predominantemente …
 6 A65     A69     Outras doenças por espiroquetas              Outras doenças por espiroquetas
 7 A70     A74     Outras doenças causadas por clamídias        Outras doenças causadas por clamídias
 8 A75     A79     Rickettsioses                                Rickettsioses
 9 A80     A89     Infecções virais do sistema nervoso central  Infecções virais do sistema nervoso central
10 A90     A99     Febres por arbovírus e febres hemorrágicas … Febres por arbovírus e febres hemorrágicas …
… with 265 more rows

}
\references{
\url{https://github.com/msrodrigues}
}
\author{
Marcio Rodrigues \email{msrodrigues@gmail.com}
}
\keyword{cid10}
\keyword{data,}
