% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cid10_document.R
\docType{data}
\name{cid_subcat}
\alias{cid_subcat}
\title{Tabela com as subcategorias do CID 10}
\description{
Trata-se de um dataset de 10.416 linhas  e 11 colinas, é a tabela mais completa do CID 10, com as
subcategorias do cid, classificação mais pormenorizada a específica de todas.
}
\details{
\itemize{
      \item cid: Código CID da subcategoria. String de 3 a 4 dígitos.
      \item classif: Classificação cruz ou asterisco.
      \item restrsexo:  indica se a categoria/subcategoria está restrita a homens (M), a mulheres (F) ou se não há restrição (ambos); o default é ambos.
      \item causaobito: Marcador se a doença pode causar óbito. Padrão é SIM
      \item descricao: Texto descrevendo a doença
      \item descricaoabrev: Texto descrevendo a doença, com o código CID de no máximo 51 caracteres
      \item refer:
      \item excluidos:
      \item cid_ord: replicação do campo subcat no formato ordered factor. Permite a comparação ordenada entre CIDs
      \item indice: indice numérico ordenado.
      \item capitulo: capitulo que o CID pertence (1 a 22)
}
}
\references{
\url{https://github.com/msrodrigues}
}
\author{
Marcio Rodrigues \email{msrodrigues@gmail.com}
}
\keyword{cid10}
\keyword{data,}
