% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cid_range.R
\name{cid_range}
\alias{cid_range}
\title{Função cid_range}
\usage{
cid_range(cat_inic, cat_final = NA, cid = FALSE)
}
\description{
Captura um range de CIDs, retornando uma tabela (data frame)
}
\details{
Argumentos:

cat_inicial: subcategoria do cid, em formato texto. Aceita receber como somente 1 letra,
uma letra e um número, ou 2 ou 3 números.
Exemplo: "A1" irá capturar todos os CIDS que começem por A1 (A1 até A199, se existirem)

cat_inf: a categoria que ira ser o final do range, no mesmo formato que a anterior.

Essa função pressupõe uma ordem nos CIDs, de A a Z e ordenada conforme os números,
na mesma ordem da lista global de cids. A cat_inicial deve preceder a cat_final

Exemplo:
cid_range("A104", "B19") retorna um range válido
cid_range("B19", "A104") retorna um erro pq a categoria inicial é posterior a categoria final.
}
\references{
\url{https://github.com/msrodrigues}
}
\author{
Marcio Rodrigues \email{msrodrigues@gmail.com}
}
\keyword{cid10}
\keyword{data,}
